Give up on intersecting constructs as a whole
After repeatedly running into inconsistencies trying to implement
efficient rule introduction on a domain where contexts overlap with
targets of preceding rules, the costs (complexity) now seem to outweigh
the benefits (universality).

In short, the combinatorial object that would faithfully represent the
space of chains of heads (overlapped sequences of nested constructions)
requires us to encode each non-construction (head termination) on the
same level as constructions which grows the sampling space of our model
in ways that don't feel necessary compared to simply combinations of
rule applications or something close to it.

Moreover, the current state of the implementation mistakenly assumes we
can ignore the whole set of suffixes of each context and have basically
three options at each inference step: i) use the whole context (append
to the head), ii) use the right/snd of the context (assuming it's
composite) (chain a new head) or iii) start a new head with an
uncorrelated atom (start a disjoint head). This assumption is wrong
based on the following counter-example: introducing a rule that
constructs a pair of symbols where the right/snd part is a composite
that chains with a subsequent head (case ii.) creates a chain that is
anchored not on the right/snd of the new construction, but on the
right/snd of the right/snd of the new construction, hence two levels
down and arbitrarily deep with generalization.

These factors make it more attractive to just move forward with a
context-free approach and avoid addressing the problem of an overly
strict alignment imposed by deriving the rule set greedily. It can
probably be solved in post-processing e.g. using an over-determining
rule set and selecting rules more carefully within that restricted
space.